+++ experimental/packages/sdk-logs/test/common/export/BatchLogRecordProcessor.test.ts
@@ -16,7 +16,7 @@ import { MeterProvider } from '@opentelemetry/sdk-metrics';
 import type {
   BatchLogRecordProcessorOptions,
   LogRecordLimits,
-  SdkLogRecord,
+  ReadWriteLogRecord,
   LogRecordExporter,
 } from '../../../src';
 import { InMemoryLogRecordExporter } from '../../../src';
@@ -37,7 +37,7 @@ class BatchLogRecordProcessor extends BatchLogRecordProcessorBase<BatchLogRecord
 const createLogRecord = (
   limits?: LogRecordLimits,
   resource?: Resource
-): SdkLogRecord => {
+): ReadWriteLogRecord => {
   const sharedState = new LoggerProviderSharedState(
     resource || defaultResource(),
     {
@@ -656,10 +656,10 @@ describe('BatchLogRecordProcessorBase', () => {
     it('should only send a single batch at a time', async () => {
       // arrange
       const callbacks: ((result: ExportResult) => void)[] = [];
-      const logRecords: SdkLogRecord[] = [];
+      const logRecords: ReadWriteLogRecord[] = [];
       const exporter: LogRecordExporter = {
         export: async (
-          exportedLogRecords: SdkLogRecord[],
+          exportedLogRecords: ReadWriteLogRecord[],
           resultCallback: (result: ExportResult) => void
         ) => {
           callbacks.push(resultCallback);